scripts/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
python-frontmatter>=1.1.0
PyYAML>=6.0
